backend/requirements.txt
# Pinned versions — generated from requirements.in
# MIN_RELEASE_AGE_DAYS = 7 (all packages below released well before April 2026)
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-telegram-bot==21.7
anthropic==0.40.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
apscheduler==3.10.4
httpx==0.28.1
aiofiles==24.1.0
pillow==11.1.0
# transitive
anyio==4.7.0
certifi==2024.12.14
google-cloud-firestore==2.19.0
google-auth==2.36.0
h11==0.14.0
httpcore==1.0.7
idna==3.10
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
